requirements.txt
Flask>=3.0,<4
pypdf>=5.0,<6
python-docx>=1.1,<2
openpyxl>=3.1,<4
pytest>=8.3,<9
